+++ src/Application/Sms/Exception/SmsException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Fight\Common\Application\Sms\Exception;
+
+use Fight\Common\Domain\Exception\SystemException;
+
+/**
+ * Class SmsException
+ */
+class SmsException extends SystemException
+{
+}